Creación y manejo de clase Categoria

+++ subastas-online-tpi/src/main/java/com/subastas/tpi/entity/Categoria.java
@@ -12,17 +12,15 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 
-@Getter
-@Setter
+@Getter @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
 @Table(name = "categorias")
-
 public class Categoria {
     
-    @Id    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
     @Column(unique = true, nullable = false)
     private String nombre;
